requirements.txt
streamlit>=1.42.0
transformers>=4.48.0
torch>=2.5.0
pillow>=10.4.0
opencv-python>=4.11.0
ultralytics>=8.4.0
accelerate>=0.33.0
sentencepiece>=0.2.0

# API and worker (Phase 1)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
celery[redis]>=5.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.12
aiohttp>=3.9.0
